# docs/swagger-backend-invillia-create-store.yaml
swagger: '2.0'
info:
  description: Spring Boot REST API for Invillia Create Store Entry
  version: '1.0'
  title: Invillia Create Store REST API
  termsOfService: Terms of service
  contact: {}
  license:
    name: Apache License Version 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8090'
basePath: /
tags:
  - name: store-controller-v-1
    description: Criação de Loja
paths:
  /api/v1/createStore:
    post:
      tags:
        - store-controller-v-1
      summary: Cria uma Loja
      operationId: createStoreUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: store
          description: store
          required: true
          schema:
            $ref: '#/definitions/Store'
      responses:
        '200':
          description: Criação de Loja realizada com sucesso
          schema:
            type: string
            example: 1234567890
        '500':
          description: Erro na criação da Loja
  /api/v1/updateStore/{id}:
    put:
      tags:
        - store-controller-v-1
      summary: Atualiza uma Loja
      operationId: updateStoreUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: store
          description: store
          required: true
          schema:
            $ref: '#/definitions/Store'
        - in: path
          name: id
          description: Store ID
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Atualização de Loja realizada com sucesso
          schema:
            type: string
            example: 1234567890
        '500':
          description: Erro na atualização da Loja
  /api/v1/findStore/{name}:
    get:
      tags:
        - store-controller-v-1
      summary: Procura uma Loja pelo Nome
      operationId: findStoreUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: name
          description: Store Name
          required: true
          type: string
      responses:
        '200':
          description: Atualização de Loja realizada com sucesso
          schema:
            type: string
            example: 'Invillia Store 0'
        '500':
          description: Erro na atualização da Loja
definitions:
  Store:
    type: object
    properties:
      name:
        type: string
        example: 'Invillia Store'
      address:
        type: string
        example: 'Rua Padre Duarte, 1253'
  Order:
    type: object
    properties:
      address:
        type: string
        example: 'Avenida XV de Novembro, 167'
      portabilityExpectedDate:
        type: string
        format: date-time
      status:
        type: string
        example: 'Pedido Realizado'
      items:
        type: array
        items:
          $ref: '#/definitions/OrderItem'
  OrderItem:
    type: object
    properties:
      description:
        type: string
        example: 'Mouse Óptico Wireless'
      unitPrice:
        type: number
        example: '70.00'
      quantity:
        type: integer
        format: int32
        example: '2'
  Payment:
    type: object
    properties:
      store:
        $ref: '#/definitions/Store'
      order:
        $ref: '#/definitions/Order'
      status:
        type: string
        example: 'Aguardando Pagamento'
      creditCardNumber:
        type: string
        example: '5421 8373 2913 8080'
      paymentDate:
        type: string
        format: date-time