# src/main/java/com/invillia/acme/store/model/OrderItem.java
package com.invillia.acme.store.model;

import java.math.BigInteger;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

@Document(collection = "orderItem")
public class OrderItem {

	@Id
	public BigInteger id;
	private String description;
	private Double unitPrice;
	private int quantity;

	public String getDescription() {
		return description;
	}

	public void setDescription(String description) {
		this.description = description;
	}

	public Double getUnitPrice() {
		return unitPrice;
	}

	public void setUnitPrice(Double unitPrice) {
		this.unitPrice = unitPrice;
	}

	public int getQuantity() {
		return quantity;
	}

	public void setQuantity(int quantity) {
		this.quantity = quantity;
	}

	@Override
	public String toString() {
		return "OrderItem [description=" + description + ", unitPrice=" + unitPrice + ", quantity=" + quantity + "]";
	}

}